requirements.txt
python-dotenv==1.0.0

Django==4.2.6
djangorestframework==3.14.0

djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9

pytest==7.4.2
pytest-django==4.5.2
coverage==7.3.2

sentry-sdk==1.32.0
